pyproject.toml
[project]
name = "ampf"
description = "Angular & Material & Python & FastAPI"
version = "0.3.05"
license = "MIT"
authors = [
    { name = "Marcin Leliwa", email = "marcin.leliwa@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0",
    "aiohttp>=3.12.15",
    "pydantic>=2.11.3",
]

[project.optional-dependencies]
fastapi = [
    "fastapi>=0.115.12",
    "gunicorn>=23.0.0",
    "pydantic-settings>=2.8.1",
    "pydantic[email]>=2.11.3",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.34.1",
]
gcp = [
    "cryptography>=44.0.2",
    "google-cloud-core>=2.4.3",
    "google-cloud-firestore>=2.20.1",
    "google-cloud-pubsub>=2.29.0",
    "google-cloud-storage>=3.1.0",
]
huggingface = [
    "sentence-transformers>=4.0.2",
]
weaviate = [
    "weaviate-client>=4.15.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "google-cloud-pubsub>=2.29.0",
    "docker>=7.1.0",
    "httpx>=0.28.1",
    "pydantic-settings>=2.9.1",
    "pydantic[email]>=2.11.5",
    "pyjwt>=2.10.1",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "pytest-cov>=6.1.1",
    "weaviate-client>=4.15.0",
    "google-cloud-firestore>=2.21.0",
    "google-cloud-core>=2.4.3",
    "google-cloud-storage>=3.1.0",
    "fastapi>=0.115.12",
    "pydantic[email]>=2.11.6",
    "pyjwt>=2.10.1",
    "pydantic-settings>=2.9.1",
    "python-multipart>=0.0.20",
    "sentence-transformers>=4.1.0",
]

[[tool.uv.index]]
name = "private-registry"
url = "https://europe-west3-python.pkg.dev/development-428212/pip"
publish-url = "https://europe-west3-python.pkg.dev/development-428212/pip"
